requirements.txt
django>=3.1.4
djangorestframework>=3.12.2
requests>=2.25.1
django-filter>=2.4.0
django-allauth>=0.44.0
asgiref>=3.3.1
django-environ==0.4.5
mysqlclient>=2.0.3
whitenoise>=5.2.0
Brotli>=1.0.9
